Create HunterAdmin.lua
I'm releasing this source, it's for my admin which I had for a long time. It's supposed to be my personal admin so.. yea keep that in mind when skidding. I use a lot of custom functions in so u can't just run it and expect it to work. I'm pretty bored of Roblox so yea I'm releasing it. Skid it to your hearts desires Idrc at this point. -Hunter (534144)